# tots_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reservation Management API",
        "description": "API for managing space reservations, allowing users to book and manage event spaces",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/spaces": {
            "get": {
                "tags": [
                    "Spaces"
                ],
                "summary": "Get all spaces",
                "description": "Retrieve a list of all available spaces with optional filters",
                "operationId": "4d73a69d5f22f345d4f45ba91c45f052",
                "parameters": [
                    {
                        "name": "min_capacity",
                        "in": "query",
                        "description": "Minimum capacity filter",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "max_capacity",
                        "in": "query",
                        "description": "Maximum capacity filter",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 100
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name or description",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Sala"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Sala A"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Sala de reuniones pequeña"
                                                    },
                                                    "capacity": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "Piso 1"
                                                    },
                                                    "image_url": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "hourly_rate": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 50
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Spaces"
                ],
                "summary": "Create new space",
                "description": "Create a new space (Admin only)",
                "operationId": "26f76a00e23873c125497f7651535f8a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "capacity",
                                    "location",
                                    "hourly_rate"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sala VIP"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Sala ejecutiva con pantalla 4K"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Piso 5"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "hourly_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 120
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Space created successfully"
                    },
                    "403": {
                        "description": "Unauthorized - Admin only"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/spaces/{id}": {
            "get": {
                "tags": [
                    "Spaces"
                ],
                "summary": "Get space by ID",
                "description": "Retrieve detailed information about a specific space",
                "operationId": "51263b3f4e4a7c14833461ffcba86833",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Space ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "capacity": {
                                                    "type": "integer"
                                                },
                                                "reservations": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Space not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Spaces"
                ],
                "summary": "Update space",
                "description": "Update an existing space (Admin only)",
                "operationId": "a2d349d3652a2f3082fc09d61880814e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "hourly_rate": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Space updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized - Admin only"
                    },
                    "404": {
                        "description": "Space not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Spaces"
                ],
                "summary": "Delete space",
                "description": "Delete an existing space (Admin only)",
                "operationId": "6b10994fdab655b7e3083432631bab0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Space deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized - Admin only"
                    },
                    "404": {
                        "description": "Space not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "http",
                "description": "Login with username and password to get the authentication token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}